ref(sdk): restructure joint apis
