+++ packages/core/src/api/index.ts
@@ -6,5 +6,11 @@ export * from './composed/spore/transferSpore';
 export * from './composed/spore/destroySpore';
 
 // Joint APIs
-export * from './joints/cluster';
-export * from './joints/spore';
+export * from './joints/cluster/injectNewClusterOutput';
+export * from './joints/cluster/injectLiveClusterCell';
+export * from './joints/cluster/injectClusterIds';
+export * from './joints/cluster/getClusterCell';
+export * from './joints/spore/injectNewSporeOutput';
+export * from './joints/spore/injectLiveSporeCell';
+export * from './joints/spore/injectSporeIds';
+export * from './joints/spore/getSporeCell';
